+++ examples/3ds_redirect.php
@@ -2,7 +2,7 @@
 <?php
 
 /**
- * PHP library version: v1.7
+ * PHP library version: 1.8
  */
 require_once('../lib/worldpay.php');
 
@@ -24,13 +24,11 @@ try {
         var_dump($response);
         echo 'There was a problem authorising 3DS order <br/>';
     }
-} catch (WorldpayException $e) { // PHP 5.3+
+} catch (WorldpayException $e) {
     // Worldpay has thrown an exception
     echo 'Error code: ' . $e->getCustomCode() . '<br/>
     HTTP status code:' . $e->getHttpStatusCode() . '<br/>
     Error description: ' . $e->getDescription()  . ' <br/>
     Error message: ' . $e->getMessage() . ' <br/>' .
     'PaRes: ' . print_r($_POST, true) . '<br/>';
-} catch (Exception $e) {  // PHP 5.2
-    echo 'Error message: '. $e->getMessage();
 }