+++ examples/header.php
@@ -28,6 +28,9 @@
 	    .form-row {
 	        height:50px;
 	    }
+	    .apm-url input {
+	    	width: 350px;
+	    }
 	    .form-row.large {
 	    	height: 150px;
 	    }
@@ -84,12 +87,12 @@
 <div class="container">
 
 <ul id="top-nav">
-    <li><a href="index.php">Create Order</a></li>&nbsp;&nbsp;|
-    <li><a href="index_cardonfile.php">Create Order (CardOnFile)</a></li>&nbsp;&nbsp;|
-    <li><a href="create_apm_form.php">Create APM Order</a></li>&nbsp;&nbsp;|
-    <li><a href="capture_authorised_form.php">Capture Authorised Order</a></li>&nbsp;&nbsp;|
-    <li><a href="cancel_authorised_form.php">Cancel Authorised Order</a></li>&nbsp;&nbsp;|
-    <li><a href="refund.php">Refund</a></li>&nbsp;&nbsp;|
-    <li><a href="partial_refund.php">Partial Refund</a></li>&nbsp;&nbsp;|
-    <li><a href="stored_cards.php">Stored Cards</a></li>
+    <li><a href="/index.php">Create Order</a></li>&nbsp;&nbsp;|
+    <li><a href="/index_cardonfile.php">Create Order (CardOnFile)</a></li>&nbsp;&nbsp;|
+    <li><a href="/capture_authorised_form.php">Capture Authorised Order</a></li>&nbsp;&nbsp;|
+    <li><a href="/cancel_authorised_form.php">Cancel Authorised Order</a></li>&nbsp;&nbsp;|
+    <li><a href="/refund.php">Refund</a></li>&nbsp;&nbsp;|
+    <li><a href="/partial_refund.php">Partial Refund</a></li>&nbsp;&nbsp;|
+    <li><a href="/stored_cards.php">Stored Cards</a></li>&nbsp;&nbsp;|
+    <li><a href="/get_order_form.php">Get Order</a></li>
 </ul>