+++ changelog.md
@@ -1,6 +1,16 @@
+### 1.8 2016-02-03
+* Shopper Language Code Fix
+* Reusable Flag for Direct Orders
+* Added CustomerIdentifiers on Cardonfile
+* Code Refactor
+* Added Missing Fields on Cardonfile Form
+* Support Direct PAN orders
+* Order Code Suffixes and Prefixes
+* Always send name as "3D" in order request for 3DS live orders
+* Empty settlement currency fix
+
 ### 1.7 2015-10-16
 * Support Zero Auth
-
 
 ### 1.6 2015-05-27
 * Check is_numeric instead of is_int for Captures